# ENH: Add faster inverse-Wishart rvs and logpdf