--- a/scipy/stats/_multivariate.py
+++ b/scipy/stats/_multivariate.py
@@ -3,14 +3,12 @@
 #
 import math
 import numpy as np
-from numpy import asarray_chkfinite, asarray
 import scipy.linalg
 from scipy._lib import doccer
 from scipy.special import (gammaln, psi, multigammaln, xlogy, entr, betaln,
                            ive, loggamma)
 from scipy._lib._util import check_random_state, _lazywhere
 from scipy.linalg.blas import drot, get_blas_funcs
-from scipy.linalg._misc import LinAlgError
 from ._continuous_distns import norm
 from ._discrete_distns import binom
 from . import _mvn, _covariance, _rcont