--- a/scipy/stats/_multivariate.py
+++ b/scipy/stats/_multivariate.py
@@ -2710,10 +2710,14 @@ class invwishart_gen(wishart_gen):
         log_det_x = np.empty(x.shape[-1])
         tr_scale_x_inv = np.empty(x.shape[-1])
         trsm = get_blas_funcs(('trsm'), (x,))
-        for i in range(x.shape[-1]):
-            Cx, log_det_x[i] = self._cholesky_logdet(x[:, :, i])
-            A = trsm(1., Cx, C, side=0, lower=True)
-            tr_scale_x_inv[i] = np.linalg.norm(A)**2
+        if dim > 1:
+            for i in range(x.shape[-1]):
+                Cx, log_det_x[i] = self._cholesky_logdet(x[:, :, i])
+                A = trsm(1., Cx, C, side=0, lower=True)
+                tr_scale_x_inv[i] = np.linalg.norm(A)**2
+        else:
+            log_det_x[:] = np.log(x[0, 0])
+            tr_scale_x_inv[:] = C[0, 0]**2 / x[0, 0]
 
         # Log PDF
         out = ((0.5 * df * log_det_scale - 0.5 * tr_scale_x_inv) -