This PR updates inverse-Wishart sampling to use the faster algorithm described in Algorithm 5 of https://arxiv.org/abs/2310.15884. It additionally updates the `logpdf` method to a faster version using the transform used in Theorem 3.2 of the same paper. 

#### Reference issue
Closes #19432 

#### What does this implement/fix?
The new `rvs` implementation eliminates one dense product, one `chol_solve`, and one `cholesky` factorization. The new `logpdf` implementation replaces an inversion via Cholesky solve, a matrix product, and a trace with a triangular solve and an l2-norm.

#### Additional information
The new `rvs` implementation was benchmarked in https://github.com/scipy/scipy/issues/19432#issuecomment-1780106406 is 1.7x faster than the old algorithm. I benchmarked the new `logpdf` implementation with the following code and saw a 2x speedup:

<details>

```python
import timeit
import numpy as np
from scipy.stats import invwishart

# Seed for reproducibility
random_seed = 42
np.random.seed(random_seed)

# Define the parameters for the inverse Wishart distribution
dim = 100  # dim
df = dim   # degrees of freedom

def rand_pd_mat(dim):
    _, R = np.linalg.qr(np.random.normal(size=(dim, dim)))
    return R.T @ R

scale = rand_pd_mat(dim)
n_samples = 10
x = np.moveaxis(invwishart.rvs(df=df, scale=scale, size=n_samples), [0, 1, 2], [2, 0, 1])

def logpdfs():
    invwishart.logpdf(x, df, scale)

# Time the function
number_of_evals = 10
total_time = timeit.repeat(logpdfs, repeat=100, number=number_of_evals)
min_time_per_run = min(total_time) / number_of_evals
print(f"min time per run: {min_time_per_run}")
```
old: 0.012830740000936203
new: 0.0061246897996170445

</details> 

cc @ilayn, since this involves some LAPACK calls.