@ilayn 2023-10-28
@WarrenWeckesser @rkern you might be interested in the sampling methodology here too (if you have any bandwidth that is).