@sethaxen 2023-10-29
I set up a benchmark using asv, and here are the results with the latest commit:

<details>
<summary> Benchmark code </summary>

```python
import numpy as np
import scipy.stats as stats
from .common import Benchmark

class InverseWishart(Benchmark):
    param_names = ["dim", "size"]
    params = [
        [1, 5, 10, 100, 1_000],
        [1, 10],
    ]

    @staticmethod
    def random_pd_mat(dim, random_state):
        A = random_state.normal(size=(dim, dim))
        _, R = np.linalg.qr(A)
        return R.T @ R

    def setup(self, dim, size):
        self.rng = np.random.default_rng(12345678)
        self.scale = self.random_pd_mat(dim, self.rng)
        self.df = dim
        self.dist_frozen = stats.invwishart(self.df, self.scale)
        self.x = self.dist_frozen.rvs(size, random_state=self.rng)
        if self.x.ndim > 2:
            self.x = np.moveaxis(self.x, [0, 1, 2], [2, 0, 1])
        self.rng = np.random.default_rng(12345678)
        if dim <= 100:
            self.number = 100

    def time_rvs(self, dim, size):
        stats.invwishart.rvs(self.df, self.scale, size=size, random_state=self.rng)

    def time_rvs_frozen(self, dim, size):
        self.dist_frozen.rvs(size, random_state=self.rng)

    def time_logpdf(self, dim, size):
        stats.invwishart.logpdf(self.x, self.df, self.scale)

    def time_logpdf_frozen(self, dim, size):
        self.dist_frozen.logpdf(self.x)
```

</details>

<details>
<summary> Benchmark results </summary>

| Change   | Before [d7f12d38] <main>   | After [ff4f389d] <invwishart_faster>   |   Ratio | Benchmark (Parameter)                             |
|----------|----------------------------|----------------------------------------|---------|---------------------------------------------------|
| -        | 68.3±0.2μs                 | 61.6±1μs                               |    0.9  | stats.InverseWishart.time_rvs_frozen(1, 1)        |
| -        | 77.1±0.6μs                 | 65.4±0.9μs                             |    0.85 | stats.InverseWishart.time_rvs_frozen(5, 1)        |
| -        | 114±0.7μs                  | 96.1±1μs                               |    0.84 | stats.InverseWishart.time_rvs_frozen(5, 10)       |
| -        | 53.1±0.5μs                 | 41.4±2μs                               |    0.78 | stats.InverseWishart.time_logpdf(1, 1)            |
| -        | 136±1μs                    | 106±1μs                                |    0.78 | stats.InverseWishart.time_rvs_frozen(10, 10)      |
| -        | 86.9±0.8μs                 | 67.1±1μs                               |    0.77 | stats.InverseWishart.time_rvs_frozen(10, 1)       |
| -        | 140±1μs                    | 105±2μs                                |    0.75 | stats.InverseWishart.time_rvs(5, 10)              |
| -        | 96.7±0.7μs                 | 71.3±0.6μs                             |    0.74 | stats.InverseWishart.time_rvs(1, 1)               |
| -        | 103±1μs                    | 75.7±0.8μs                             |    0.73 | stats.InverseWishart.time_rvs(5, 1)               |
| -        | 93.5±1μs                   | 67.7±1μs                               |    0.72 | stats.InverseWishart.time_rvs_frozen(1, 10)       |
| -        | 165±1μs                    | 118±2μs                                |    0.71 | stats.InverseWishart.time_rvs(10, 10)             |
| -        | 2.74±0.02ms                | 1.87±0.05ms                            |    0.68 | stats.InverseWishart.time_rvs_frozen(100, 10)     |
| -        | 117±0.7μs                  | 76.9±0.7μs                             |    0.66 | stats.InverseWishart.time_rvs(10, 1)              |
| -        | 123±2μs                    | 78.2±0.8μs                             |    0.64 | stats.InverseWishart.time_rvs(1, 10)              |
| -        | 3.03±0.03ms                | 1.95±0.05ms                            |    0.64 | stats.InverseWishart.time_rvs(100, 10)            |
| -        | 279±2μs                    | 172±2μs                                |    0.62 | stats.InverseWishart.time_logpdf(10, 10)          |
| -        | 254±5μs                    | 156±1μs                                |    0.62 | stats.InverseWishart.time_logpdf(5, 10)           |
| -        | 46.2±0.3μs                 | 28.8±0.7μs                             |    0.62 | stats.InverseWishart.time_logpdf_frozen(1, 1)     |
| -        | 267±0.9μs                  | 163±2μs                                |    0.61 | stats.InverseWishart.time_logpdf_frozen(10, 10)   |
| -        | 244±3μs                    | 150±0.7μs                              |    0.61 | stats.InverseWishart.time_logpdf_frozen(5, 10)    |
| -        | 1.03±0s                    | 614±8ms                                |    0.6  | stats.InverseWishart.time_rvs_frozen(1000, 10)    |
| -        | 472±1μs                    | 280±0.7μs                              |    0.59 | stats.InverseWishart.time_logpdf(100, 1)          |
| -        | 105±1ms                    | 60.6±0.8ms                             |    0.58 | stats.InverseWishart.time_rvs_frozen(1000, 1)     |
| -        | 1.12±0s                    | 627±5ms                                |    0.56 | stats.InverseWishart.time_rvs(1000, 10)           |
| -        | 105±0.5ms                  | 57.9±0.7ms                             |    0.55 | stats.InverseWishart.time_logpdf(1000, 1)         |
| -        | 2.73±0.01ms                | 1.49±0.01ms                            |    0.54 | stats.InverseWishart.time_logpdf(100, 10)         |
| -        | 423±2μs                    | 229±0.8μs                              |    0.54 | stats.InverseWishart.time_logpdf_frozen(100, 1)   |
| -        | 2.68±0.01ms                | 1.43±0ms                               |    0.53 | stats.InverseWishart.time_logpdf_frozen(100, 10)  |
| -        | 502±3μs                    | 265±5μs                                |    0.53 | stats.InverseWishart.time_rvs_frozen(100, 1)      |
| -        | 115±0.5μs                  | 58.3±0.4μs                             |    0.51 | stats.InverseWishart.time_logpdf(10, 1)           |
| -        | 921±5ms                    | 453±0.9ms                              |    0.49 | stats.InverseWishart.time_logpdf(1000, 10)        |
| -        | 108±0.7μs                  | 52.6±1μs                               |    0.49 | stats.InverseWishart.time_logpdf(5, 1)            |
| -        | 91.2±1ms                   | 44.2±0.6ms                             |    0.48 | stats.InverseWishart.time_logpdf_frozen(1000, 1)  |
| -        | 909±2ms                    | 440±3ms                                |    0.48 | stats.InverseWishart.time_logpdf_frozen(1000, 10) |
| -        | 106±0.4μs                  | 47.7±0.5μs                             |    0.45 | stats.InverseWishart.time_logpdf_frozen(10, 1)    |
| -        | 101±2μs                    | 43.5±1μs                               |    0.43 | stats.InverseWishart.time_logpdf_frozen(5, 1)     |
| -        | 769±10μs                   | 314±9μs                                |    0.41 | stats.InverseWishart.time_rvs(100, 1)             |
| -        | 199±1ms                    | 74.2±0.3ms                             |    0.37 | stats.InverseWishart.time_rvs(1000, 1)            |
| -        | 137±0.4μs                  | 35.7±0.7μs                             |    0.26 | stats.InverseWishart.time_logpdf(1, 10)           |
| -        | 134±2μs                    | 23.1±0.5μs                             |    0.17 | stats.InverseWishart.time_logpdf_frozen(1, 10)    |

</details>